+++ src/hooks/commons/index.ts
@@ -1,0 +1,1 @@
+export { default as useScroll } from './useScroll';
